move DSPlotAccessor to accessor module
